Type cast return value to int
Ceil() returns a float, but the task clearly stated that expected value
should be of type integer.

+++ src/TimeComplexity/FrogJump.php
@@ -6,6 +6,6 @@ class FrogJump
 {
     public function jumpInIntervalFromStartToDestination(int $start, int $destination, int $jumpLength)
     {
-        return ceil(($destination - $start) / $jumpLength);
+        return (int) ceil(($destination - $start) / $jumpLength);
     }
 }
